requirements.txt
mysql-connector-python
flask
flask_login
werkzeug
Flask-Session